Code refactoring, unnecessary properties removed - HookContainer moved to the HookInput.cs
